refactor: Rule type validation (#3400)
* refactor: UserInfo validation


* refactor: Rule type validation
