+++ api/kyverno/v1/rule_test.go
@@ -1,0 +1,152 @@
+package v1
+
+import (
+	"encoding/json"
+	"testing"
+
+	"gotest.tools/assert"
+	"k8s.io/apimachinery/pkg/util/validation/field"
+)
+
+func Test_Validate_RuleType_EmptyRule(t *testing.T) {
+	subject := Rule{
+		Name: "validate-user-privilege",
+	}
+	path := field.NewPath("dummy")
+	errs := subject.Validate(path)
+	assert.Equal(t, len(errs), 1)
+	assert.Equal(t, errs[0].Field, "dummy")
+	assert.Equal(t, errs[0].Type, field.ErrorTypeInvalid)
+	assert.Equal(t, errs[0].Detail, "No operation defined in the rule 'validate-user-privilege'.(supported operations: mutate,validate,generate,verifyImages)")
+
+}
+
+func Test_Validate_RuleType_MultipleRule(t *testing.T) {
+	rawPolicy := []byte(`
+	{
+		"spec": {
+		   "rules": [
+			  {
+				 "name": "validate-user-privilege",
+				 "match": {
+					"resources": {
+					   "kinds": [
+						  "Deployment"
+					   ],
+					   "selector": {
+						  "matchLabels": {
+							 "app.type": "prod"
+						  }
+					   }
+					}
+				 },
+				 "mutate": {
+					"patchStrategicMerge": {
+					   "spec": {
+						  "template": {
+							 "spec": {
+								"containers": [
+								   {
+									  "(name)": "*",
+									  "resources": {
+										 "limits": {
+											"+(memory)": "300Mi",
+											"+(cpu)": "100"
+										 }
+									  }
+								   }
+								]
+							 }
+						  }
+					   }
+					}
+				 },
+				 "validate": {
+					"message": "validate container security contexts",
+					"anyPattern": [
+					   {
+						  "spec": {
+							 "template": {
+								"spec": {
+								   "containers": [
+									  {
+										 "securityContext": {
+											"runAsNonRoot": true
+										 }
+									  }
+								   ]
+								}
+							 }
+						  }
+					   }
+					]
+				 }
+			  }
+		   ]
+		}
+	 }`)
+
+	var policy *ClusterPolicy
+	err := json.Unmarshal(rawPolicy, &policy)
+	assert.NilError(t, err)
+	for _, rule := range policy.Spec.GetRules() {
+		path := field.NewPath("dummy")
+		errs := rule.Validate(path)
+		assert.Assert(t, len(errs) != 0)
+	}
+}
+
+func Test_Validate_RuleType_SingleRule(t *testing.T) {
+	rawPolicy := []byte(`
+	{
+		"spec": {
+		   "rules": [
+			  {
+				 "name": "validate-user-privilege",
+				 "match": {
+					"resources": {
+					   "kinds": [
+						  "Deployment"
+					   ],
+					   "selector": {
+						  "matchLabels": {
+							 "app.type": "prod"
+						  }
+					   }
+					}
+				 },
+				 "validate": {
+					"message": "validate container security contexts",
+					"anyPattern": [
+					   {
+						  "spec": {
+							 "template": {
+								"spec": {
+								   "containers": [
+									  {
+										 "securityContext": {
+											"runAsNonRoot": "true"
+										 }
+									  }
+								   ]
+								}
+							 }
+						  }
+					   }
+					]
+				 }
+			  }
+		   ]
+		}
+	 }
+	`)
+
+	var policy *ClusterPolicy
+	err := json.Unmarshal(rawPolicy, &policy)
+	assert.NilError(t, err)
+	for _, rule := range policy.Spec.GetRules() {
+		path := field.NewPath("dummy")
+		errs := rule.Validate(path)
+		assert.Assert(t, len(errs) == 0)
+	}
+}
