+++ api/kyverno/v1/rule_types.go
@@ -1,0 +1,137 @@
+package v1
+
+import (
+	"fmt"
+	"reflect"
+
+	"k8s.io/apiextensions-apiserver/pkg/apis/apiextensions"
+	apiextv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
+	"k8s.io/apimachinery/pkg/util/validation/field"
+)
+
+// Rule defines a validation, mutation, or generation control for matching resources.
+// Each rules contains a match declaration to select resources, and an optional exclude
+// declaration to specify which resources to exclude.
+type Rule struct {
+	// Name is a label to identify the rule, It must be unique within the policy.
+	// +kubebuilder:validation:MaxLength=63
+	Name string `json:"name,omitempty" yaml:"name,omitempty"`
+
+	// Context defines variables and data sources that can be used during rule execution.
+	// +optional
+	Context []ContextEntry `json:"context,omitempty" yaml:"context,omitempty"`
+
+	// MatchResources defines when this policy rule should be applied. The match
+	// criteria can include resource information (e.g. kind, name, namespace, labels)
+	// and admission review request information like the user name or role.
+	// At least one kind is required.
+	MatchResources MatchResources `json:"match,omitempty" yaml:"match,omitempty"`
+
+	// ExcludeResources defines when this policy rule should not be applied. The exclude
+	// criteria can include resource information (e.g. kind, name, namespace, labels)
+	// and admission review request information like the name or role.
+	// +optional
+	ExcludeResources ExcludeResources `json:"exclude,omitempty" yaml:"exclude,omitempty"`
+
+	// Preconditions are used to determine if a policy rule should be applied by evaluating a
+	// set of conditions. The declaration can contain nested `any` or `all` statements. A direct list
+	// of conditions (without `any` or `all` statements is supported for backwards compatibility but
+	// will be deprecated in the next major release.
+	// See: https://kyverno.io/docs/writing-policies/preconditions/
+	// +optional
+	RawAnyAllConditions *apiextv1.JSON `json:"preconditions,omitempty" yaml:"preconditions,omitempty"`
+
+	// Mutation is used to modify matching resources.
+	// +optional
+	Mutation Mutation `json:"mutate,omitempty" yaml:"mutate,omitempty"`
+
+	// Validation is used to validate matching resources.
+	// +optional
+	Validation Validation `json:"validate,omitempty" yaml:"validate,omitempty"`
+
+	// Generation is used to create new resources.
+	// +optional
+	Generation Generation `json:"generate,omitempty" yaml:"generate,omitempty"`
+
+	// VerifyImages is used to verify image signatures and mutate them to add a digest
+	// +optional
+	VerifyImages []*ImageVerification `json:"verifyImages,omitempty" yaml:"verifyImages,omitempty"`
+}
+
+// HasMutate checks for mutate rule
+func (r *Rule) HasMutate() bool {
+	return !reflect.DeepEqual(r.Mutation, Mutation{})
+}
+
+// HasVerifyImages checks for verifyImages rule
+func (r *Rule) HasVerifyImages() bool {
+	return r.VerifyImages != nil && !reflect.DeepEqual(r.VerifyImages, ImageVerification{})
+}
+
+// HasValidate checks for validate rule
+func (r *Rule) HasValidate() bool {
+	return !reflect.DeepEqual(r.Validation, Validation{})
+}
+
+// HasGenerate checks for generate rule
+func (r *Rule) HasGenerate() bool {
+	return !reflect.DeepEqual(r.Generation, Generation{})
+}
+
+// MatchKinds returns a slice of all kinds to match
+func (r *Rule) MatchKinds() []string {
+	matchKinds := r.MatchResources.ResourceDescription.Kinds
+	for _, value := range r.MatchResources.All {
+		matchKinds = append(matchKinds, value.ResourceDescription.Kinds...)
+	}
+	for _, value := range r.MatchResources.Any {
+		matchKinds = append(matchKinds, value.ResourceDescription.Kinds...)
+	}
+
+	return matchKinds
+}
+
+// ExcludeKinds returns a slice of all kinds to exclude
+func (r *Rule) ExcludeKinds() []string {
+	excludeKinds := r.ExcludeResources.ResourceDescription.Kinds
+	for _, value := range r.ExcludeResources.All {
+		excludeKinds = append(excludeKinds, value.ResourceDescription.Kinds...)
+	}
+	for _, value := range r.ExcludeResources.Any {
+		excludeKinds = append(excludeKinds, value.ResourceDescription.Kinds...)
+	}
+	return excludeKinds
+}
+
+func (r *Rule) GetAnyAllConditions() apiextensions.JSON {
+	return FromJSON(r.RawAnyAllConditions)
+}
+
+func (r *Rule) SetAnyAllConditions(in apiextensions.JSON) {
+	r.RawAnyAllConditions = ToJSON(in)
+}
+
+// ValidateRuleType checks only one type of rule is defined per rule
+func (r *Rule) ValidateRuleType(path *field.Path) field.ErrorList {
+	var errs field.ErrorList
+	ruleTypes := []bool{r.HasMutate(), r.HasValidate(), r.HasGenerate(), r.HasVerifyImages()}
+	count := 0
+	for _, v := range ruleTypes {
+		if v {
+			count++
+		}
+	}
+	if count == 0 {
+		errs = append(errs, field.Invalid(path, r, fmt.Sprintf("No operation defined in the rule '%s'.(supported operations: mutate,validate,generate,verifyImages)", r.Name)))
+	} else if count != 1 {
+		errs = append(errs, field.Invalid(path, r, fmt.Sprintf("Multiple operations defined in the rule '%s', only one operation (mutate,validate,generate,verifyImages) is allowed per rule", r.Name)))
+	}
+	return errs
+}
+
+// Validate implements programmatic validation
+func (r *Rule) Validate(path *field.Path) field.ErrorList {
+	var errs field.ErrorList
+	errs = append(errs, r.ValidateRuleType(path)...)
+	return errs
+}
